changed focus and break time variables to unsigned 16 bit integers.

+++ include/Apps/FocusAssist/focus_assist_controller.h
@@ -11,8 +11,8 @@ private:
         count++;
     }
 
-    uint8_t focusTimeSec = 45 * 60;
-    uint8_t breakTimeSec = 15 * 60;
+    uint16_t focusTimeSec = 45 * 60;
+    uint16_t breakTimeSec = 15 * 60;
 
 public:
     FocusAssistController() {}
